Updated build system
+ Author and author email
+ add to setup.py that setuptools is required.
urgent: had to comment out Devons unittest.py file, I tried rebuilding to make sure my changes to setup.py didn't screw anything up. Since the test file is a newly added file and the imports are messed up it caused the setup.py not to build properly. **This is a good reminder for us, so when sunday comes we can have our changes finalized and we can rerun setup.py to make sure it works. **