New code for the highscore task
database updates with the score, name, and game_id. I am not attaching the CLI code that I changed because I do not want to cause A merge conflict, so I will post the file in roughdraft code. The CLI giveup function currently prints out what should be getting sent to the database only as a test to make sure it is actually sending the right info, it will not be a permanent part of the code.