Highscore fix
tweaked some of the naming and comments (may have missed some so if I did just please point them out). All commented out code from me should be removed. Keeping in some test code to aid with debugging for now, but nothing is test code in the model, only in the CLI (which I'm not actually pushing yet) and the Highscores.py file