Fixes for highscore
Added the highscores.db to the .gitignore file so that it remains a local db. it will still get created on everyones machine when running the code, this just prevents us pushing individual versions of the db, which will be very different depending on the puzzles we play. Also, added the code back into the connectDB.py because Dr. Schwartz said I needed to add the connection in this file, and it did not work when I removed it like was originally requested, so if it something that 100% needs removed then we will need to find a fix for it.