One line in model
Replaced a missing line from an earlier push. Calls the generateGameID in the gameLoad function. All this does is create an identical ID for the puzzle that was saved and then loaded. I am not saving the gameID to a json file, instead it associates with the database when it is generated to check if high scores exist.