CLI Score
CLI now automatically checks if the puzzle is finished and automatically calls the giveup function when it is finished. I updated the giveUp function and it is a little more complex now, as it has a conditional argument that needed to be passed into the invoker that checks whether the puzzle is finished. Because I thought it would be redundant to ask the player if they wanted to give up when they finished, but didn't want to make a new method. So if the puzzle isn't finished and you call giveUp, it will prompt you yes/no, but upon submitting the final word it will skip the yes/no and just ask you for your username instead. This took a while to implement but I think the pay off is really good.