Score GUI Update
Implements the required client fixes in GUI. High scores button is now in file, and giving up reverts to the empty game state. Created a new function resetGUI to help with that. Added an instance of the 6 buttons for the regular letters to allow them to be destroyed when giving up. This also implements the "auto giveup" when you guess every word, and just like the cli, I made it so that if you are giving up regularly then it will prompt you yes/no when giving up, but if you guess every word then it skips the yes/no and just asks for your username as it would otherwise be redundant.